adapt to renaming repo as hark
